Course List App
This is a course list app that I designed using JSON Server. When the page loads, it displays a list of available courses, each with an 'Enquire' button. When the user clicks on this button, they are navigated to another tab where they can enter their enquiry and submit it. All submitted queries are displayed in the Enquiries tab.